+++ react/MutationAddToCart.js
@@ -1,0 +1,3 @@
+import addToCart from './mutations/addToCart.graphql'
+
+export default addToCart