+++ react/MutationInsertCoupon.js
@@ -1,0 +1,3 @@
+import insertCoupon from './mutations/insertCoupon.graphql'
+
+export default insertCoupon