+++ react/Queries.tsx
@@ -1,3 +1,6 @@
+/** DEPRECATED, being kept just for retrocompatibility.
+ * Create individual entry points for each query instead */
+
 import orderForm from './queries/orderForm.graphql'
 import cardSessionId from './queries/cardSessionId.graphql'