+++ react/QueryCardSessionId.js
@@ -1,0 +1,3 @@
+import cardSessionId from './queries/cardSessionId.graphql'
+
+export default cardSessionId