+++ react/QueryOrderForm.js
@@ -1,0 +1,3 @@
+import orderForm from './queries/orderForm.graphql'
+
+export default orderForm